blackpink-auto-clicker-1/requirements.txt
PyQt5
discord.py
pyscreeze
pyperclip
keyboard
psutil
pywin32